Added tests and some refactor

+++ src/main/java/com/project/paymentology/apis/dtos/TransactionReconcileDto.java
@@ -15,7 +15,7 @@ public class TransactionReconcileDto {
 
     private LocalDateTime transactionDate;
 
-    private double transactionAmount;
+    private long transactionAmount;
 
     private String transactionNarrative;
